requirements.txt
numpy
soundfile
librosa
python-dotenv
openai
google-cloud-pubsub
